refactor image_viewer_text_field
