+++ lib/settings/utils/save_image_viewer_app.dart
@@ -1,0 +1,6 @@
+import 'package:shared_preferences/shared_preferences.dart';
+
+Future<void> saveImageViewerApp(String value) async {
+  final prefs = await SharedPreferences.getInstance();
+  await prefs.setString('imageViewerApp', value);
+}
